LayoutTests/fast/css/flex-calculated-value-expected.txt

Tests assigning a calculated values to 'flex' CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['flex'] is ""
testDiv.style['flex'] = 'calc(2 * 3) calc(2 * 3)'
PASS testDiv.style['flex'] is "6 6 0%"
PASS window.getComputedStyle(testDiv).getPropertyValue('flex') is "6 6 0%"
testDiv.style['flex'] = 'calc(2 * 3) calc(2 * 3) calc(10% * 2)'
PASS testDiv.style['flex'] is "6 6 calc(20%)"
PASS window.getComputedStyle(testDiv).getPropertyValue('flex') is "6 6 20%"
testDiv.style['flex'] = 'calc(2 * 3) calc(2 * 3) calc(3 - 3)'
PASS testDiv.style['flex'] is "6 6 0px"
PASS window.getComputedStyle(testDiv).getPropertyValue('flex') is "6 6 0px"
PASS successfullyParsed is true

TEST COMPLETE
